Revert "adding the DTO to commuction with user and entity"
