+++ src/main/java/com/note/Notes/Entity/User.java
@@ -1,0 +1,34 @@
+package com.note.Notes.Entity;
+
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import jakarta.persistence.*;
+import lombok.*;
+
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+
+     @Column(nullable=false,unique = true)
+    private String userName;
+
+    @Column(nullable=false)
+     private String password;
+    @Column(nullable=false,unique=true)
+    private String userEmail;
+
+    @OneToMany( mappedBy = "userId", cascade = CascadeType.ALL,orphanRemoval = true)
+    @JsonManagedReference
+      List<Notes> notesList=new ArrayList<>();
+
+}
