+++ src/main/java/com/note/Notes/Exceptions/DuplicateEmail.java
@@ -1,0 +1,8 @@
+package com.note.Notes.Exceptions;
+
+public class DuplicateEmail extends RuntimeException{
+    public DuplicateEmail(String message)
+    {
+          super(message);
+    }
+}
