+++ src/main/java/com/note/Notes/Exceptions/DuplicateTitle.java
@@ -1,0 +1,8 @@
+package com.note.Notes.Exceptions;
+
+public class DuplicateTitle extends RuntimeException{
+    public DuplicateTitle(String message)
+    {
+         super(message);
+    }
+}
