+++ src/main/java/com/note/Notes/Exceptions/DuplicateUser.java
@@ -1,0 +1,8 @@
+package com.note.Notes.Exceptions;
+
+public class DuplicateUser extends RuntimeException{
+    public DuplicateUser(String message)
+    {
+        super(message);
+    }
+}
