+++ src/main/java/com/note/Notes/Exceptions/InternalServerErrorException.java
@@ -1,0 +1,12 @@
+package com.note.Notes.Exceptions;
+
+public class InternalServerErrorException extends  RuntimeException{
+    public InternalServerErrorException(String message) {
+        super(message);
+    }
+
+    public InternalServerErrorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
