+++ src/main/java/com/note/Notes/Exceptions/NotesNotFoundException.java
@@ -1,0 +1,10 @@
+package com.note.Notes.Exceptions;
+
+public class NotesNotFoundException extends RuntimeException {
+
+     public NotesNotFoundException(String message)
+     {
+
+          super(message);
+     }
+}
