+++ src/main/java/com/note/Notes/Exceptions/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.note.Notes.Exceptions;
+
+public class UserNotFoundException extends RuntimeException {
+
+     public UserNotFoundException(String message)
+     {
+          super(message);
+     }
+}
