+++ src/main/java/com/note/Notes/NotesApplication.java
@@ -1,0 +1,13 @@
+package com.note.Notes;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotesApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NotesApplication.class, args);
+	}
+
+}
