+++ src/main/java/com/note/Notes/Repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.note.Notes.Repository;
+
+import com.note.Notes.Entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+    Optional<User> findByUserName(String userName);
+    Optional<User> findByUserEmail(String email);
+}
