--- src/main/java/org/repo/NotesApplication.java
@@ -1,13 +1,0 @@
-package org.repo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotesApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NotesApplication.class, args);
-	}
-
-}
