--- src/main/java/org/repo/configuration/ConfiSecurity.java
@@ -1,50 +1,0 @@
-package org.repo.configuration;
-
-import org.repo.service.CustomUserService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-
-@Configuration
-@EnableWebSecurity
-public class ConfiSecurity {
-
-    private final CustomUserService customUserService;
-
-    public ConfiSecurity(CustomUserService customUserService) {
-        this.customUserService = customUserService;
-    }
-   @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .csrf(csrf->csrf.disable())
-                .authorizeHttpRequests(
-                        auth->auth
-                                .requestMatchers("/register").permitAll()
-                                .anyRequest().authenticated())
-                     .httpBasic(Customizer.withDefaults());
-
-      return httpSecurity.build();
-  }
-
-  @Bean
-  public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception
-  {
-      return authenticationConfiguration.getAuthenticationManager();
-  }
-
-  @Bean
-    public PasswordEncoder passwordEncoder(){
-      return new BCryptPasswordEncoder();
-  }
-}
-
