--- src/main/java/org/repo/dto/NotesRequestDTO.java
@@ -1,17 +1,0 @@
-package org.repo.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class NotesRequestDTO {
-    private long id;
-    private String title;
-    private String content;
-
-}
