--- src/main/java/org/repo/dto/UserRequestDTO.java
@@ -1,16 +1,0 @@
-package org.repo.dto;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserRequestDTO {
-    private long id;
-    private String userName;
-    private String password;
-    private String UserEmail;
-}
