--- src/main/java/org/repo/dto/UserResponseDTO.java
@@ -1,18 +1,0 @@
-package org.repo.dto;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserResponseDTO {
-    private long id;
-    private String userName;
-    private String userEmail;
-    private List<NotesResponseDTO> listOfNotes;
-}
