--- src/main/java/org/repo/exceptions/DuplicateEmail.java
@@ -1,8 +1,0 @@
-package org.repo.exceptions;
-
-public class DuplicateEmail extends RuntimeException{
-    public DuplicateEmail(String message)
-    {
-          super(message);
-    }
-}
