--- src/main/java/org/repo/exceptions/DuplicateTitle.java
@@ -1,8 +1,0 @@
-package org.repo.exceptions;
-
-public class DuplicateTitle extends RuntimeException{
-    public DuplicateTitle(String message)
-    {
-         super(message);
-    }
-}
