--- src/main/java/org/repo/exceptions/DuplicateUser.java
@@ -1,8 +1,0 @@
-package org.repo.exceptions;
-
-public class DuplicateUser extends RuntimeException{
-    public DuplicateUser(String message)
-    {
-        super(message);
-    }
-}
