--- src/main/java/org/repo/exceptions/InternalServerErrorException.java
@@ -1,12 +1,0 @@
-package org.repo.exceptions;
-
-public class InternalServerErrorException extends  RuntimeException{
-    public InternalServerErrorException(String message) {
-        super(message);
-    }
-
-    public InternalServerErrorException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
