--- src/main/java/org/repo/exceptions/NotesNotFoundException.java
@@ -1,10 +1,0 @@
-package org.repo.exceptions;
-
-public class NotesNotFoundException extends RuntimeException {
-
-     public NotesNotFoundException(String message)
-     {
-
-          super(message);
-     }
-}
