--- src/main/java/org/repo/exceptions/UserNotFoundException.java
@@ -1,9 +1,0 @@
-package org.repo.exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-
-     public UserNotFoundException(String message)
-     {
-          super(message);
-     }
-}
