--- src/main/java/org/repo/repository/NoteRepository.java
@@ -1,18 +1,0 @@
-package org.repo.repository;
-
-import org.repo.entity.Notes;
-import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface NoteRepository extends JpaRepository<Notes,Long> {
-    @Transactional
-    @Modifying
-      void  deleteByTitle(String title);
-   Optional<Notes> findByTitle(String title);
-
-}
