--- src/main/java/org/repo/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package org.repo.repository;
-
-import org.repo.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-    Optional<User> findByUserName(String userName);
-    Optional<User> findByUserEmail(String email);
-}
