+++ src/main/resources/application.properties
@@ -7,4 +7,3 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 spring.output.ansi.enabled=ALWAYS
 server.port=8081
-
