+++ src/test/java/com/note/Notes/NotesApplicationTests.java
@@ -1,0 +1,13 @@
+package com.note.Notes;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotesApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
