--- src/test/java/org/repo/NotesApplicationTests.java
@@ -1,13 +1,0 @@
-package org.repo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NotesApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
